fix(audit): pass 1 confirmation — StubSettings round-trips the grouping key, grouping tests split out, guard and comment fixes
Confirmation round 1 over the remediation (nothing above LOW):

- tests/unit/helpers/StubSettings.h: overrides the new ISettings defaulted
  pair with a real member and NOTIFY emit, like its two siblings, so a
  StubSettings-backed adaptor no longer drops the write into the no-op body.
- libs/phosphor-scroll-engine/tests: the six grouping tests move to a new
  test_scrollengine_grouping.cpp (510 lines, registered through
  pse_add_test with a vertical arm), bringing test_scrollengine_behaviour.cpp
  back under the ceiling (1374 to 947). columnOfOrFail's doc now says it
  only warns and every caller guards the index before .at() or an equality
  compare.
- engine_grouping.cpp: resolveTile returns by value (a reference into the
  QHash memo would die on rehash) and its comment says the memo is defensive
  now that the scan skips the active column.
- engine_overrides.cpp: drops the ScrollStrip.h include nothing uses since
  the split; effectiveBoolOverride's caller count is nine.
- ScrollEngine.h: the toggle block counts five map-only toggles plus the
  focus arm and names center-short-columns; "six toggles above".
- isettings.h / settings/scrolling.cpp: five defaulted scrolling getters,
  the grouping toggle named in both notes; settings.h says the setter
  overrides ISettings alone.
- test_scrolling_settings.cpp: QVERIFY(settings.save()).
