+++ libs/phosphor-scroll-engine/src/scrollengine/engine_overrides.cpp
@@ -15,7 +15,6 @@
 // header trim cannot break the cast (the nounity build's role in catching
 // exactly this).
 #include <PhosphorScrollEngine/IScrollSettings.h>
-#include <PhosphorScrollEngine/ScrollStrip.h>
 
 #include "enginelimits.h"
 
@@ -185,7 +184,7 @@ StripAxis ScrollEngine::effectiveStripAxis(const QVariantMap& overrides, const Q
 }
 
 // ── behaviour toggles ──
-// One shape for all eight callers (the five behaviour toggles plus the three
+// One shape for all nine callers (the six behaviour toggles plus the three
 // tab-indicator bools): a rule-written per-screen key wins, an absent key
 // falls back to the member the global config seeded. The value is taken only
 // when it is a real bool — a hand-edited string would otherwise coerce to
