+++ src/config/settings.h
@@ -1537,10 +1537,11 @@ public:
 
     // Store-backed scalars under Scrolling.Behavior; mostly shared leaf key
     // names (FocusNewWindows, StickyWindowHandling, …) disambiguated by
-    // group, plus the scrolling-only GroupSameAppAsTabs. That pair satisfies
-    // both ISettings (defaulted virtuals, so the D-Bus registry can publish
+    // group, plus the scrolling-only GroupSameAppAsTabs. Its getter satisfies
+    // both ISettings (defaulted virtual, so the D-Bus registry can publish
     // the key through the interface) and IScrollSettings (the engine's live
-    // read), hence the override on the setter too.
+    // read) with one body; its setter overrides the ISettings no-op alone,
+    // hence the override on it too.
     bool scrollingFocusNewWindows() const override;
     void setScrollingFocusNewWindows(bool focus);
     bool scrollingGroupSameAppAsTabs() const override;
